refactor(rbac): rename template policy permission tables
Refs: HOL-994
